+++ public/admin/_app/immutable/nodes/24.COMhMz1E.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/BcFzXtc-.js";export{e as component};